+++ src/App.tsx
@@ -1,6 +1,6 @@
 import { Outlet } from "react-router-dom";
 import { createGlobalStyle } from "styled-components";
-import Navbar from "./components/layout/navbar/navbar";
+import Navbar from "./components/layout/navbar/LayoutNavbar";
 import { AuthProvider } from "./contexts/AuthContext";
 
 const GlobalStyle = createGlobalStyle`
